Refactor http module into submodules.
